+++ test/HBProtocol.test.js
@@ -150,7 +150,6 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           expect(await this.dsRegistry.proxies.call(someone)).eq(ZERO_ADDRESS);
 
           const to = this.hMaker.address;
-          const value = ether('5');
           const ilkEth = utils.padRight(utils.asciiToHex('ETH-A'), 64);
           // const wadD = ether('500');
           const [
@@ -158,6 +157,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
             minCollateral,
           ] = await getGenerateLimitAndMinCollateral(ilkEth);
           const wadD = generateLimit;
+          const value = minCollateral;
           const data = abi.simpleEncode(
             'openLockETHAndDraw(uint256,address,address,bytes32,uint256)',
             value,
@@ -168,7 +168,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           );
           const receipt = await this.proxy.execMock(to, data, {
             from: someone,
-            value: ether('10'),
+            value: value,
           });
           const handlerReturn = utils.toBN(
             getHandlerReturn(receipt, ['uint256'])[0]
@@ -201,13 +201,13 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           expect(await this.dsRegistry.proxies.call(someone)).eq(ZERO_ADDRESS);
 
           const to = this.hMaker.address;
-          const value = ether('5');
           const ilkEth = utils.padRight(utils.asciiToHex('ETH-A'), 64);
           const [
             generateLimit,
             minCollateral,
           ] = await getGenerateLimitAndMinCollateral(ilkEth);
           const wadD = generateLimit;
+          const value = minCollateral;
           const data = abi.simpleEncode(
             'openLockETHAndDraw(uint256,address,address,bytes32,uint256)',
             MAX_UINT256,
@@ -244,13 +244,13 @@ contract('BProtocol', function([_, user1, user2, someone]) {
 
         it('User has proxy', async function() {
           const to = this.hMaker.address;
-          const value = ether('5');
           const ilkEth = utils.padRight(utils.asciiToHex('ETH-A'), 64);
           const [
             generateLimit,
             minCollateral,
           ] = await getGenerateLimitAndMinCollateral(ilkEth);
           const wadD = generateLimit;
+          const value = minCollateral;
           const data = abi.simpleEncode(
             'openLockETHAndDraw(uint256,address,address,bytes32,uint256)',
             value,
@@ -261,7 +261,7 @@ contract('BProtocol', function([_, user1, user2, someone]) {
           );
           const receipt = await this.proxy.execMock(to, data, {
             from: user1,
-            value: ether('10'),
+            value: value,
           });
           const handlerReturn = utils.toBN(
             getHandlerReturn(receipt, ['uint256'])[0]
@@ -286,13 +286,13 @@ contract('BProtocol', function([_, user1, user2, someone]) {
 
         it('User has proxy with max amount', async function() {
           const to = this.hMaker.address;
-          const value = ether('5');
           const ilkEth = utils.padRight(utils.asciiToHex('ETH-A'), 64);
           const [
             generateLimit,
             minCollateral,
           ] = await getGenerateLimitAndMinCollateral(ilkEth);
           const wadD = generateLimit;
+          const value = minCollateral;
           const data = abi.simpleEncode(
             'openLockETHAndDraw(uint256,address,address,bytes32,uint256)',
             MAX_UINT256,